requirements.txt
#bot
python-telegram-bot==12.8
google-cloud-pubsub==2.9.0

# microservices integration
google-api-python-client==2.34.0
google-auth==2.33

google-cloud-logging==2.7

#storing
pandas==1.3.5
google-cloud-storage==2.0.0
google-cloud-firestore

# yarl is a workaround according to https://issuetracker.google.com/issues/162329927
# yarl==1.4.2
